requirements.txt
pycryptodome
gmpy2
ecdsa